report: split topbar features
