+++ report/test/renderer/report-ui-features-test.js
@@ -342,7 +342,7 @@ describe('ReportUIFeatures', () => {
       window = dom.document().defaultView;
       const features = new ReportUIFeatures(dom);
       features.initFeatures(sampleResults);
-      dropDown = features._dropDown;
+      dropDown = features._topbar._dropDownMenu;
     });
 
     it('click should toggle active class', () => {
